Restore compatibility with Contao 4.8
Until now the alias of the backported translator was always set because
the container builder in the extension only has access to it's own
service definitions and not to all one. That's why the alias definition
has to move to a compiler pass.

It also changes when the cca backport translator decorates the symfony
translator: It now only happens if the Contao translator is not
available.